fix: link list common funcs created | DRY
