+++ python3/Competitive programming/recursion/link-list/reverse.py
